Be able to turn off the spanned text for custom notification layout (#1)
Added a flag to the builder whether or not the spanned text should be used for the custom notification layout. For some cases we want the spanned text to be used only for the regular notification style, without propagation to the custom one, as it might define a totally different style of the text.